# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Jewellery ERP API
  version: 0.1.0
  description: >
    Phase 1 API for single-store jewellery ERP. Gold loans, interest logic,
    and multi-branch flows are explicitly excluded.
servers:
  - url: http://localhost:4000
paths:
...
  /auth/session:
    post:
      tags: [auth]
      summary: Verify OTP and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyRequest"
      responses:
        "200":
          description: Session created
  /inventory/items:
    get:
      tags: [inventory]
      summary: List items
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Item list
    post:
      tags: [inventory]
      summary: Create item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
      responses:
        "201":
          description: Item created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OtpVerifyRequest:
      type: object
      required: [phone, otp]
      properties:
        phone:
          type: string
        otp:
          type: string
    CreateItemRequest:
      type: object
      required: [itemCode, category, metal, grossWt, netWt, purity]
      properties:
        itemCode:
          type: string
        category:
          type: string
        metal:
          type: string
          enum: [GOLD, SILVER, DIAMOND]
        grossWt:
          type: number
        netWt:
          type: number
        purity:
          type: number
        makingCharges:
          type: number
        rateAtPurchase:
          type: number
...